+++ src/layout/Align/types.ts
@@ -1,0 +1,13 @@
+export enum AlignMethod {
+  CENTER = 'center',
+  CENTER_X = 'centerX',
+  CENTER_Y = 'centerY',
+  BOTTOM_IN = 'bottomIn',
+  BOTTOM_TO = 'bottomTo',
+  TOP_IN = 'topIn',
+  TOP_TO = 'topTo',
+  RIGHT_IN = 'rightIn',
+  RIGHT_TO = 'rightTo',
+  LEFT_IN = 'leftIn',
+  LEFT_TO = 'leftTo'
+}