+++ src/layout/Layout/Layout.ts
@@ -1,0 +1,16 @@
+import { LayoutArea } from "../LayoutArea/types";
+
+interface IProps {
+  layoutArea: LayoutArea;
+}
+
+
+/** Возвращает методы для макетирования **/
+export class Layout {
+    constructor(private props: IProps) {
+    }
+
+    sidebar(){
+
+    }
+}