+++ src/layout/LayoutArea/types.ts
@@ -1,0 +1,4 @@
+export type LayoutArea = Phaser.Scene | Phaser.Geom.Rectangle | { getBounds: () => Phaser.Geom.Rectangle } | {
+  width: number,
+  height: number
+};