+++ storybook/stories.ts
@@ -4,8 +4,9 @@ import { launchSceneAutoRunAndPauseMainStory } from "../src/scenes/ChildScene/la
 import { launchSceneAutoRunStory } from "../src/scenes/ChildScene/launchSceneAutoRun.story";
 import { launchSceneStory2 } from "../src/scenes/ChildScene/launchScene2.story";
 import { launchSceneStory } from "../src/scenes/ChildScene/launchScene.story";
-import { alignStory } from "../src/layout/Align.story";
+import { alignStory } from "../src/layout/Align/Align.story";
 import { viewPortStory } from "../src/components/ViewPort/ViewPort.story";
+import { alignChainStory } from "../src/layout/alignChains/alignChainStory";
 
 export const stories: IStoryListItem[] = Array.from(
   new Set<IStoryListItem>([
@@ -13,6 +14,7 @@ export const stories: IStoryListItem[] = Array.from(
     viewPortStory,
     storyTitle("Align"),
     alignStory,
+    alignChainStory,
     storyTitle("Child Scene"),
     launchSceneAutoRunAndPauseMainStory,
     launchSceneAutoRunStory,